% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_analysis.r
\name{band_analysis}
\alias{band_analysis}
\title{Frequency band analysis}
\usage{
band_analysis(df, freq.bands, log_transform = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe of predicted smoothed power spectrum estimated from spectrum_est()}

\item{freq.bands}{matrix specifying lower and upper bound of frequency band (e.g. (t(c(0, 2/24)), or rbind(c(0,2/24), c(2/24,25), c(25, 60))). Bounds must be between 0 and 60)}

\item{log_transform}{specifies if log power spectrum should be calculated. TRUE is default}
}
\description{
Internal function used to estimate area under power spectrum at particular frequency bands
}
\author{
Jessica Graves
}
