% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_predicted.R
\docType{data}
\name{df_predicted}
\alias{df_predicted}
\title{Example dataframe generated from for RAR_ExCosine()}
\format{A data frame with 23820 rows and 12 variables (note, there are 60 missing time steps in rar_data3, and so nrows(df_predicted) is 60 fewer than rar_data3). Activity measurements are acquired for every minute (60-second epochs). Number of observation days = 18}
\usage{
data(df_predicted)
}
\description{
df_predicted.rda is an example dataframe that is generated from RAR_ExCosine(). In general, dataframes generated from RAR_ExCosine() can be accessed using RAR_ExCosine()$.
}
\details{
Variables include:
\itemize{
  \item time: factor listing time of observation.
  \item act: observed activity (raw counts).
  \item log.act: log(act + 1) generated by RAR_ExCosine()
  \item day: day column generated by RAR_ExCosine()
  \item time.posixct: POSIXct time formate generated by RAR_ExCosine()
  \item tm: time in minutes from midnight
  \item td: time in minutes from the first midnight
  \item n: observation number
  \item predicted: predicated values estimated from Extended Cosine model, on log scale
  \item resid: calculated residauls estimated from Extended Cosine model, on log scale
  \item e_predicted: predicted values on natural scale
  \item e_resid: residuals on natural scale
}
}
\keyword{datasets}
