% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localized.r
\name{localized}
\alias{localized}
\title{Localized}
\usage{
localized(df, log_transform = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe containing actigraphy data and time.}

\item{log_transform}{accepts logical specifying if log-transformation should be applied to activity data}
}
\description{
Internal function used to analyze mean, standard deviation and relative activity within specified time-bins
}
\author{
Jessica Graves
}
