% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pable.R
\name{Pable}
\alias{Pable}
\title{Table formating}
\usage{
Pable(table, Head = F, n = 5, scale = F, Format = "markdown")
}
\description{
Function for creating tables for kniting in a nice format
 @param table name of the table object
 @param Head if you want to show only the first observations, default is F
 @param n if Head=T specify the number of rows to show, default is 5
 @param scale scale to fit the margins, use when number of variables is large
 @keywords Table
 @examples
 Pable(Table)
}
