% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person_time.r
\name{person_time}
\alias{person_time}
\title{Person Time}
\usage{
person_time(df, hour_bin, wake_hr)
}
\arguments{
\item{df}{dataframe containing actigraphy data and time.}

\item{hour_bin}{specifies length of time-bin.}

\item{wake_hr}{specifies the average wake hour (can be number or column in dataset if multiple subjects).}
}
\description{
Internal function used to adjust for average waking hour before conducting localized RAR analysis
}
\author{
Jessica Graves
}
