% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCBivarPlot.R
\name{PROC.Bivar.Plot}
\alias{PROC.Bivar.Plot}
\title{Bivariate plots}
\usage{
PROC.Bivar.Plot(Exposure, Outcome, Data, alpha = 0.05)
}
\arguments{
\item{Exposure}{Exposure of interest, must be a string of characters}

\item{Outcome}{Outcome of interest, must be a string of characters}

\item{Data}{Dataset to use}

\item{alpha}{0.05 by default}
}
\description{
Function for creating informative bivariate plots
}
