% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcBivar.R
\name{PROC.Bivar}
\alias{PROC.Bivar}
\title{Bivariate analysis for catevorical vs Categorical}
\usage{
PROC.Bivar(Outcome, Exposure, Data, alpha = 0.05)
}
\arguments{
\item{Outcome}{Outcome of interest, must be a string of characters}

\item{Exposure}{Exposure of interest, must be a string of characters}

\item{Data}{Dataset that includes the variables}

\item{alpha}{0.05 by default}
}
\description{
Function for 2x2 tables
}
