% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCBivariateRegression.R
\name{PROC.BivariateRegression}
\alias{PROC.BivariateRegression}
\title{Multiple Bivariate regressions table}
\usage{
PROC.BivariateRegression(variables, outcome, data, alpha = 0.05,
  family = "gaussian", Offset = Offset)
}
\arguments{
\item{variables}{a vector of variable names to be analyzed}

\item{data}{name of the data that contains the variables}

\item{alpha}{0.05 by default}

\item{family}{"gaussian" by default, specifies the type of regression. Currently supports "gaussian", "binomial", and "poisson"}

\item{Offset}{if family = "poisson", include the name of the variable that will be used as offset.}

\item{Outcome}{the outcome of interest to be analyzed}
}
\description{
This function provides a table of the estimates of a bivariate regression (either linear or logistic) with the exposure of interest
}
