% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCChisquTbl.R
\name{PROC.ChisqTbl}
\alias{PROC.ChisqTbl}
\title{Table for multiple chi square tests}
\usage{
PROC.ChisqTbl(variables, outcome, Data, alpha = 0.05)
}
\arguments{
\item{variables}{a vector of variable names to be analyzed}

\item{alpha}{0.05 by default}

\item{Outcome}{the outcome of interest to be analyzed}

\item{data}{name of the data that contains the variables}
}
\description{
This function provides a table with multiple chi suare tests results
}
