% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCplot.R
\name{PROC.plot}
\alias{PROC.plot}
\title{Function for continuous variables plots.}
\usage{
PROC.plot(variable)
}
\arguments{
\item{variable}{variable to use, must be continous}
}
\description{
This function provides visualization for the distribution of a continous variable
}
