% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCplott.R
\name{PROC.plot.t}
\alias{PROC.plot.t}
\title{Function for ploting the transformations}
\usage{
PROC.plot.t(variable)
}
\arguments{
\item{variable}{name of the variable to transform, must be continous}
}
\description{
This functions plots 4 posible transformations
}
