% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCtransformations.R
\name{PROC.transformations}
\alias{PROC.transformations}
\title{Function for transforming the data:}
\usage{
PROC.transformations(variable)
}
\arguments{
\item{variable}{name of the variable to transform}
}
\description{
Provides a table with some of the parameters for the transformations made (log10, ln, squrt and sqr)
}
