% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCunivarL.R
\name{PROC.univar.L}
\alias{PROC.univar.L}
\title{Univariate for a list of variables}
\usage{
PROC.univar.L(variables, data)
}
\arguments{
\item{variables}{must be a string of the variables name as on the data}

\item{data}{name of the data from where the variables are from}
}
\description{
Provides a summary descriptibe table for a list of variables
}
