% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCunivar.R
\name{PROC.univar}
\alias{PROC.univar}
\title{Function for summary tables}
\usage{
PROC.univar(variable, is.categorical = F)
}
\description{
Results a summary descriptive table for a variable, could be either continous or categorical
}
