% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.r
\name{processing}
\alias{processing}
\title{Data processing}
\usage{
processing(df)
}
\arguments{
\item{df}{dataframe containing actigraphy data and time.}
}
\description{
Internal function used to process activity data files to work in RAR functions.
}
\author{
Jessica Graves
}
