% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR.r
\name{RAR}
\alias{RAR}
\title{RAR: Extended Cosine Model}
\usage{
RAR(df, act_column, time_column, transform = c("antilogit", "arctan",
  "hill"), id_column = NULL)
}
\arguments{
\item{df}{dataframe containing actigraphy data and time.}

\item{act_column}{name of the column within df that contains the activity count data.}

\item{time_column}{name of the column that contains date and time of observation. Time must be a POSIX object.}

\item{transform}{specifies which transformation to apply. Options include Hill Function ("hill"), Anti-Logistic ("antilogit"), or Arctangent ("arctan")}

\item{id_column}{name of column containing id if multiple subjects exist in dataframe.}
}
\description{
This function fits sigmoidally transformed extended cosine model to activity data, as seen in  Marler et al. (2006).
}
\details{
Outputs from this function include: coefficient estimates for baseline cosine model and user-specificed extended cosine transformation, predicted values, and parameter estimates of interest.
}
\examples{
data(age_wise)
d <- age_wise[age_wise$id==1,]
rar_ex <- RAR(d, act, date_time)
rar_ex$parameters # parameter estimates
rar_ex$messages # convergence message

# Multiple subjects
d4 <- age_wise[age_wise$id \%in\% c(1:4), ]
rar_ex4 <- RAR(d4, act, date_time, id_column=id)
rar_ex4$parameters # parameter estimates
rar_ex4$messages # convergence messages for each participant
rar_ex4$df_predicted # dataframe of obseved activity and predicted values
rar_ex4$df_interp # predictions based on interpolated data (i.e. missing data)
}
\references{
1. Marler M.R., Gehrman P., Martin J.L., Ancoli-Israel S. (2006) The sigmoidally transformed cosine curve: a mathematical model for circadian rhythms with symmetric non-sinusoidal shapes. Stat Med. Nov 30;25(22):3893-904.
}
\seealso{
\code{nls}
}
\author{
Jessica Graves
}
