% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rar_data.R
\docType{data}
\name{rar_data}
\alias{rar_data}
\title{Example Data for RAR_ExCosine()}
\format{A data frame with 22905 rows and 2 variables. Activity measurements are acquired for every minute (60-second epochs). Number of observation days = 19.}
\usage{
data(rar_data)
}
\description{
rar_data.rda is an example dataframe illustrating the basic format RAR_ExCosine() will accept. Note that a day column is not needed.
}
\details{
Variables include:
\itemize{
  \item time: factor listing time of observation.
  \item act: observed activity (raw counts).
}
}
\keyword{datasets}
