% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rar_data2.R
\docType{data}
\name{rar_data2}
\alias{rar_data2}
\title{Example Data 2 for RAR_ExCosine()}
\format{A data frame with 27393 rows and 2 variables. Activity measurements are acquired for every minute (60-second epochs). Number of observation days = 20.}
\usage{
data(rar_data2)
}
\description{
rar_data2.rda is another example dataframe illustrating the basic format RAR_ExCosine() will accept. Note that a day column is not needed.
}
\details{
Variables include:
\itemize{
  \item time: factor listing time of observation.
  \item act: observed activity (raw counts).
}
}
\keyword{datasets}
