% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_ExCosine.R
\name{RAR_ExCosine}
\alias{RAR_ExCosine}
\title{RAR Extended Cosine Model}
\usage{
RAR_ExCosine(df, act_column, time_column, transform = c("hill",
  "antilogit", "arctan"), plot = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe containing actigraphy data and time. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{act_column}{specifies the name of the column within df that contains the activity count data. RAR_ExCosine will do a log(activity + 1) transformation.}

\item{time_column}{specifies the name of the column that contains time of observation}

\item{transform}{specifies which transformation to apply. Options include Hill Function ("hill"), Anti-Logistic ("antilogit"), or Arctangent ("arctan")}

\item{plot}{logical specifying if plots should be outputted. If missing, defaults to FALSE.}
}
\description{
This function fits sigmoidally transformed extended cosine model to activity data, as seen in  Marler et al. (2006). Activity data provided to the function should include data for only one subject. Function takes dataframe with columns that include time and raw activity counts.
}
\details{
Outputs from this function include: coefficient estimates for baseline cosine model and user-specificed extended cosline model; starting values for baseline cosine model; dataframe with predicted values based on fitted user-specified extended cosine; and parameter estimates of interest.
}
\examples{
data(rar_data)
rar_cosine = RAR_ExCosine(rar_data, act, time, "antilogit", plot=TRUE)
rar_cosine$final_parameters
rar_cosine$plot_log.act
}
\references{
1. Marler M.R., Gehrman P., Martin J.L., Ancoli-Israel S. (2006) The sigmoidally transformed cosine curve: a mathematical model for circadian rhythms with symmetric non-sinusoidal shapes. Stat Med. Nov 30;25(22):3893-904.
}
\seealso{
\code{nls} \code{dplyr}
}
\author{
Jessica Graves
}
