% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_ExCosine_Multi.R
\name{RAR_ExCosine_Multi}
\alias{RAR_ExCosine_Multi}
\title{RAR Extended Cosine Model: For Files with Mulitple Participants}
\usage{
RAR_ExCosine_Multi(df, act_column, time_column, transform = c("hill",
  "antilogit", "arctan"), id_column, plot = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe containing actigraphy data, time, and ID column. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{act_column}{specifies the name of the column within df that contains the activity count data. RAR_ExCosine will do a log(activity + 1) transformation.}

\item{time_column}{specifies the name of the column that contains time of observation}

\item{transform}{specifies which transformation to apply. Options include Hill Function ("hill"), Anti-Logistic ("antilogit"), or Arctangent ("arctan")}

\item{id_column}{specifies the name of the column that contains ID information.}

\item{plot}{logical specifying if plots should be outputted. If missing, defaults to FALSE.}
}
\description{
This function peforms the same functionality as RAR_ExCosine(), excpet that it accepts a dataframe with a third column specifying ID to allow for iteration across subjects.
}
\details{
Outputs from this function include a tibble object containing RAR_ExCosine_Out for each participant.
}
\examples{
data(rar_data_multi)
rar_multi = RAR_ExCosine_Multi(rar_data_multi, act, time, "antilogit", id, plot=TRUE)
rar_multi$RAR_Ouput # tibble object of participants and objects

rar_multi$RAR_Output$RAR_ExCosine_Out[[1]]$final_parameters # individ participant estimates
rar_multi$RAR_Output$RAR_ExCosine_Out[[1]]$plot_log.act # individ participant plots

rar_multi$final_parameters # df of parameter estimates for all participants
}
\references{
1. Marler M.R., Gehrman P., Martin J.L., Ancoli-Israel S. (2006) The sigmoidally transformed cosine curve: a mathematical model for circadian rhythms with symmetric non-sinusoidal shapes. Stat Med. Nov 30;25(22):3893-904.
}
\seealso{
\code{nls} \code{dplyr}
}
\author{
Jessica Graves
}
