% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_FFTbands.R
\name{RAR_FFTbands}
\alias{RAR_FFTbands}
\title{RAR Fast Fourier Transform of Selected Frequency Bands}
\usage{
RAR_FFTbands(df, resid_column, time_column, predict_column,
  plot = c(TRUE, FALSE), freq.bands, sampling.rate)
}
\arguments{
\item{df}{dataframe including residuals calculating from extended cosine model based on actigraphy data. Residuals must be in numeric class. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{resid_column}{specfies the name of the column in the df that contains residuals. This must be in numeric class}

\item{time_column}{specifies the name of the column that contains time of observation for residuals}

\item{predict_column}{specifies the name of the column that contains the predicted values from the extended cosine model}

\item{plot}{a logical value specifying if plots should be outputted. Plots illustrate effects of filters.}

\item{freq.bands}{a kx2 matrix or dataframe specifcing frequency band intervals of interested for calculating area under spectrum, where the 1st column is the lower bound, and the 2nd column is the upper bound for the frequency band. Frequencies supplied must be in cycles/hour.}

\item{sampling.rate}{a numeric value representing the sampling rate of data in Hz (cycles/second). e.g. data collected once every 30 seconds specified as sampling.rate=1/30.}
}
\value{
a list containing original dataframe along with filter results, plots of filter effects (if plot=T), and lists of FFT per filter.
}
\description{
This function performs Fast Fourier Transform (FFT) on residuals based on extended cosine model (RAR_ExCosine()) and applies user-specified frequency-band filters. The methodology can be found in detail in R.T Krafty et al, Measuring Variability in Rest-Activity Rhythms from Actigraphy with Application to Characterizing Symptoms of Depression (2019)
}
\details{
This function will apply user-specified frequency bands to residuals extracted from the extened cosine model and will illustrate the effect of those filters.
}
\examples{
data(df_predicted)
f1 <- c(0, 2/24) # Filter Values
f2 <- c(2/24, 25)
f3 <- c(25, 60)
fs = as.data.frame(rbind(f1, f2, f3))
filter = RAR_FFTbands(df_predicted, resid, time, predicted, TRUE, fs, 1/60)
filter$plots[[1]]
}
\references{
1. R.T Krafty et al, Measuring Variability in Rest-Activity Rhythms from Actigraphy with Application to Characterizing Symptoms of Depression (2019)
}
\seealso{
\code{fft}
}
\author{
Jessica Graves
}
