% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_Local_plot.r
\name{RAR_Local_plot}
\alias{RAR_Local_plot}
\title{RAR_Local_plot}
\usage{
RAR_Local_plot(rar_local_object, id_vals = NULL)
}
\arguments{
\item{rar_local_object}{default output from RAR_Local().}

\item{id_vals}{character vector to print plots for each participant specified. Default is NULL. Use "ALL" to see activity measures for all participants simultaneously.}
}
\description{
Visualizes observed and fitted activity data based on mean, standard deviation, and relative activity based on user-specified time bins
}
\examples{
data(age_wise)
d <- age_wise[age_wise$id==1,]
local <- RAR_Local(d, act, date_time, 4, TRUE, 6.17) # get localized measures
RAR_Local_plot(local)

# Multiple subjects
d4 <- age_wise[age_wise$id \%in\% c(1:4), ]
rar_ex4 <- RAR(d4, act, date_time, id_column=id)
d4.2 <- merge(d4, rar_ex4$parameters, by="id")
local4 <- RAR_Local(d4.2, act, date_time, 4, wake_hr=tLeft, id_column=id) # wake-hour adjusted
RAR_Local_plot(local4, id_vals=c("1", "2"))

local4.clock <- RAR_Local(d4.2, act, date_time, 4, id_column=id) # clock time, log scale
RAR_Local_plot(local4.clock, id_vals=c("1", "2"))
}
\author{
Jessica Graves
}
