% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_Localized.R
\name{RAR_Localized}
\alias{RAR_Localized}
\title{RAR Localized Measures}
\usage{
RAR_Localized(df, act_column, time_column, hour_bin,
  person_time = c(TRUE, FALSE), tLeft = NULL, transform = c(TRUE,
  FALSE), plot = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe containing actigraphy data and time. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{act_column}{specifies the name of the column within df that contains the activity count data. RAR_ExCosine will do a log(activity + 1) transformation.}

\item{time_column}{specifies the name of the column that contains time of observation}

\item{hour_bin}{specifices the number of hours within each time bin (e.g. 4 hour timebins = 4)}

\item{person_time}{logical specifying if person-specific time adjustements should be made. If missing, default is FALSE.}

\item{tLeft}{numeric value indicating subject's up-mesor (or wake up hour). Can be estimated using RAR_ExCosine(). Or, if merged into dataframe, is column name.}

\item{transform}{logical specifying if log(activity + 1) transformation is desired}

\item{plot}{logical specifying if plots should be outputted. If missing, defaults to FALSE.}
}
\description{
This function calculates localized measures of RAR timing, as described in Graves (2018). Measures include: Mean Activity, Standard Deviation of Activity, and Relative Activity at user-specified time-bins across days.
}
\details{
Outputs from this function include: dataframes of calculated localized measures based on user-specified hour bins, day x time summaries of activity, plots of Mean, SD, and Relative Activity Measures.
}
\examples{
data(rar_data)
# defaults to clock-time and log(act + 1)
rar_local = RAR_Localized(rar_data, act, time, 4, plot=TRUE)
rar_local$localized_measures
rar_local$plots
rar_local$plots[[1]]
}
\author{
Jessica Graves
}
