% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_Localized_Multi.R
\name{RAR_Localized_Multi}
\alias{RAR_Localized_Multi}
\title{RAR Localized Measures: For Files with Mulitple Participants}
\usage{
RAR_Localized_Multi(df, act_column, time_column, id_column, hour_bin,
  person_time = c(TRUE, FALSE), tLeft = NULL, transform = c(TRUE,
  FALSE), plot = c(TRUE, FALSE))
}
\arguments{
\item{df}{dataframe containing actigraphy data, time, an ID column, and a column for average wake up hours. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{act_column}{specifies the name of the column within df that contains the activity count data. RAR_ExCosine will do a log(activity + 1) transformation.}

\item{time_column}{specifies the name of the column that contains time of observation}

\item{id_column}{specifies the name of the column that contains participant IDs}

\item{hour_bin}{specifices the number of hours within each time bin (e.g. 4 hour timebins = 4)}

\item{person_time}{logical specifying if person-specific time adjustements should be made. If missing, default is FALSE.}

\item{tLeft}{is a column name}

\item{transform}{logical specifying if log(activity + 1) transformation is desired}

\item{plot}{logical specifying if plots should be outputted. If missing, defaults to FALSE.}
}
\description{
This function calculates localized measures of RAR timing, as described in Graves (2018). Measures include: Mean Activity, Standard Deviation of Activity, and Relative Activity at user-specified time-bins across days. This function takes a single dataframe including activity, time, ID, and wake-up hour for each participant.
}
\details{
Outputs from this function include: dataframes of calculated localized measures based on user-specified hour bins, day x time summaries of activity, plots of Mean, SD, and Relative Activity Measures.
}
\examples{
data(rar_data_multi_wake)
Local_Multi = RAR_Localized_Multi(rar_data_multi_wake, act, time, id, 4, TRUE, wakehr, TRUE, TRUE)
Local_Multi$localized_measures
Local_Multi$plots
}
\author{
Jessica Graves
}
