% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_Spectrum.R
\name{RAR_Spectrum}
\alias{RAR_Spectrum}
\title{RAR Residuals spectrum analysis and smoothing splines}
\usage{
RAR_Spectrum(df, resid_column, time_column, plot = c(TRUE, FALSE),
  method = c("smooth spline", "penalized spline"), log = c(TRUE,
  FALSE), freq.bands = NULL, ...)
}
\arguments{
\item{df}{dataframe including residuals calculating from extended cosine model based on actigraphy data. Residuals must be in numeric class. Time must be in HH:MM:SS format and stored as a character, factor, or POSIX object.}

\item{resid_column}{specfies the name of the column in the df that contains residuals. This must be in numeric class}

\item{time_column}{specifies the name of the column that contains time of observation for residuals}

\item{plot}{a logical value specifying if plots should be outputted}

\item{method}{specifies the method using for the smoothing spline. method="smooth spline" indicates using a cubic smoothing spline to the supplied data. method="penalized spline" indicates using a penalized likelihood regression smoothing spline method}

\item{log}{a logical value specifying if log-transfromation should be used for the power spectrum}

\item{freq.bands}{a kx2 matrix or dataframe specifcing frequency band intervals of interested for calculating area under spectrum, where the 1st column is the lower bound, and the 2nd column is the upper bound for the frequency band. Frequencies supplied must be in cycles/hour.}

\item{...}{additional arguments passed from astsa::mvspec, stats::smooth.spline,gss::gssanova}
}
\value{
a list containg residuals and time, power spectrum and corresponding frequency, plots list and corresponding methods if plot argument is set to be true
}
\description{
This function performs Residual Circadian Spectrum (RCS) analysis from rest-activity rhythm actigraphy data. The methodology can be found in detail in R.T Krafty et al, Measuring Variability in Rest-Activity Rhythms from Actigraphy with Application to Characterizing Symptoms of Depression (2019)
}
\details{
This function will perform power spectrum analysis for residuals extracting from the extended cosine model. Plots will be draw with power spectrum lines and the smoothing curve. Log argument indicates whether log-transformation will be performed for spectrum. The function includes two methods for smoothing spline: cubic smoothing spline and penalized likelihood regression smoothing spline.
}
\examples{
data(df_predicted)
spec = RAR_Spectrum(df_predicted, resid, time, TRUE, "smooth spline", TRUE) # Using smooth spline and no frequency band analysis
spec$plot_log_ss
f1 <- c(0, 2/24) # Filter Values
f2 <- c(2/24, 25)
f3 <- c(25, 60)
fs = as.data.frame(rbind(f1, f2, f3))
spec2 = RAR_Spectrum(df_predicted, resid, time, TRUE, "smooth spline", TRUE, fs)
spec2$freq.bands # to get scores
}
\references{
1. R.T Krafty et al, Measuring Variability in Rest-Activity Rhythms from Actigraphy with Application to Characterizing Symptoms of Depression (2019)

2. Chong Gu Smoothing Spline ANOVA Models

3. R.H Shumway, D.S Stoffer, Time Series Analysis and Its Applications
}
\seealso{
\code{mvspec} \code{gssanova} \code{smooth.spline}
}
\author{
Haoyi Fu
}
