% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.r
\name{transformations}
\alias{transformations}
\title{Transformations}
\usage{
transformations(df, transform = c("antilogit", "arctan", "hill"))
}
\arguments{
\item{df}{dataframe containing actigraphy data and time.}

\item{transform}{specifies which transformation to use}
}
\description{
Internal function used to fit sigmoidally transformed extended cosine model. As
}
\author{
Jessica Graves
}
